+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All significant changes to this project will be documented in this file.
 
+## [2.0.0](https://github.com/tonystone/tracelog-adaptive-writer/tree/2.0.0)
+
+## Changed
+- Changed AdaptiveWriter to conform to TraceLog 5.0.0 Writer protocol.
+- Changed to compile with Swift 5.
+- Changed minimum deployment target for OSX to 10.13.
+- Change to us CocoaPods 1.8.4 for release.
+
 ## [1.0.0](https://github.com/tonystone/tracelog-adaptive-writer/tree/1.0.0)
 
 Initial Public Release.